requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
decorator==4.4.2
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
moviepy==1.0.3
narwhals==2.10.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
proglog==0.1.12
protobuf==6.33.0
pyarrow==22.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
